separated out Balance into folder of 3 files, per team lead suggestions
