+++ client/src/pages/Balance/index.js
@@ -1,0 +1,3 @@
+import Balance from "./Balance";
+
+export default Balance;
